+++ web-client-18LF326-XC8.X/modules/transmitter/transmitter.h
@@ -10,7 +10,7 @@
 
 void transmitter_init(void(*onReady)(void));
 uint8_t transmitter_sendNext(void);
-uint8_t transmitter_send(uint8_t * data, uint8_t length);
+uint8_t transmitter_send(uint8_t * data, uint16_t length);
 uint8_t transmitter_isBusy(void);
 
 #endif	/* TRANSMITTER_H */
